+++ CHANGES.rst
@@ -91,6 +91,8 @@ Changelog
 - For field `MeetingCategory.category_mapping_when_cloning_to_other_mc`, display
   also disabled categories in vocabulary so it is visible on category view.
   [gbastien]
+- Completed `IEncodeSecretVotes.validate_votes` to ensure values are integers.
+  [gbastien]
 
 4.2b13 (2021-07-16)
 -------------------